+++ src/lib/components/issue-card.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { Node } from '../../global';
   import Card from './card.svelte';
-  import Outline from './outline.svelte';
+  import Label from './label.svelte';
 
   export let issue: Node;
 </script>
@@ -28,7 +28,7 @@
   </div>
   <div class="mt-2 space-y-2">
     {#each issue.labels.edges as label}
-      <Outline text={label.node.name} color={label.node.color} />
+      <Label text={label.node.name} color={label.node.color} />
     {/each}
   </div>
 </Card>
